% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog.R
\name{get_catalog}
\alias{get_catalog}
\title{Get the OGD catalog}
\usage{
get_catalog()
}
\value{
A dataframe with meta information on the OGD data catalog of the Cantone of Thurgau
}
\description{
Function to get meta information on the current catalog of datasets available on data.tg.ch
}
\examples{
get_catalog()

}
