% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Get dataset by ID}
\usage{
get_dataset(dataset_id)
}
\arguments{
\item{dataset_id}{ID of the dataset from the catalog}
}
\value{
dataframe containing the dataset with the corresponding ID from data.tg.ch
}
\description{
Function to get specific dataset by ID from the datasets available on data.tg.ch
}
\examples{
get_dataset("sk-stat-4")

}
