% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_field_description.R
\name{get_field_description}
\alias{get_field_description}
\title{Get metadata by ID}
\usage{
get_field_description(dataset_id)
}
\arguments{
\item{dataset_id}{ID of the dataset from the catalog}
}
\value{
dataframe containing the metadata of the dataset with the corresponding ID from data.tg.ch
}
\description{
Function to get specific metadata by ID from the datasets available on data.tg.ch
}
\examples{
get_field_description("sk-stat-4")

}
