% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_data.R
\name{get_geo_data}
\alias{get_geo_data}
\title{Get Geo Data}
\usage{
get_geo_data(dataset_id, feature = NULL)
}
\arguments{
\item{dataset_id}{dataset_id from the catalog ( \code{\link[tgAPI]{getCatalog}})}

\item{feature}{the name or the title of a feature. If the feature is NULL, the first feature in the feature table will be used. Provide "all" to receive all features in a list of data.frames.}
}
\value{
a data.frame or a list of data.frames usable with sf
}
\description{
Retrieves the data for the given dataset_id and returns a data.frame. It is recommended to pass the feature you want to get data on as an argument (see see \code{\link[tgAPI]{get_geo_features}}.
}
