% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_feature_slice.R
\name{get_geo_feature_slice}
\alias{get_geo_feature_slice}
\title{Get Slice of Feature}
\usage{
get_geo_feature_slice(startindex, count_num, wfs, feature, bbox = NULL)
}
\arguments{
\item{startindex}{index of the record to start}

\item{wfs}{the WFSClient}

\item{feature}{the feature name}

\item{bbox}{a character string of the bbox to resttrict the request to a specific geographic region (for example: `"2670694.24166528787463903,1243765.72999201575294137,2774916.42190200183540583,1287552.7291286657564342"`)}

\item{count}{number of records that should be downloaded}
}
\value{
data.frame of class `sf`
}
\description{
Downloads a specific number of feature data
}
