% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_feature_slice_vec.R
\name{get_geo_feature_slice_vec}
\alias{get_geo_feature_slice_vec}
\title{Get Geo Data in Slices (vectorized)}
\usage{
get_geo_feature_slice_vec(num_fts, count_num, wfs, feature, bbox = NULL)
}
\arguments{
\item{num_fts}{num of features (see `get_geo_features_num()`)}

\item{count_num}{number of records that should be downloaded}

\item{wfs}{the WFSClient}

\item{feature}{the feature name}

\item{bbox}{a character string of the bbox to resttrict the request to a specific geographic region (for example: `"2670694.24166528787463903,1243765.72999201575294137,2774916.42190200183540583,1287552.7291286657564342"`)}
}
\value{
a data.frame of class `sf`
}
\description{
Downloads packages of data in slices and binds it to one data.frame
}
