% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_features.R
\name{get_geo_features}
\alias{get_geo_features}
\title{Get Geo Features}
\usage{
get_geo_features(dataset_id)
}
\arguments{
\item{dataset_id}{dataset_id from the catalog (see \code{\link[tgAPI]{getCatalog}})}
}
\value{
data.frame with the available features (layers)
}
\description{
Returns a data.frame with the available features (layers) for the given dataset_id
}
