% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_features_num.R
\name{get_geo_features_num}
\alias{get_geo_features_num}
\title{Get number of Features}
\usage{
get_geo_features_num(wfs, feature, bbox = NULL)
}
\arguments{
\item{wfs}{the WFSClient}

\item{feature}{name of the feature (for example ms:LCOBJ)}

\item{bbox}{a character string of the bbox to resttrict the request to a specific geographic region (for example: `"2670694.24166528787463903,1243765.72999201575294137,2774916.42190200183540583,1287552.7291286657564342"`)}
}
\value{
number of features
}
\description{
Returns the number of Features to estimate the loading time
}
