% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_features_wfs.R
\name{get_geo_features_wfs}
\alias{get_geo_features_wfs}
\title{Get Geo Features and WFS}
\usage{
get_geo_features_wfs(dataset_id)
}
\arguments{
\item{dataset_id}{dataset_id from the catalog (see \code{\link[tgAPI]{getCatalog}})}
}
\value{
a list containing the WFSClient and a data.frame with the available features (layers)
}
\description{
Returns a list containing the WFSClient and a data.frame with the available features (layers) for the given dataset_id
}
