% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get detailed Meta Data (only for internal usage with API Key)}
\usage{
get_metadata(key)
}
\arguments{
\item{key}{API key from OpenDataSoft}
}
\value{
dataframe containing the meta data.
}
\description{
Function to retrieve meta data from the backend. This function is for internal use only. An API key must be specified and passed to the function.
}
