% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_from_opendataswiss.R
\name{get_url_from_opendataswiss}
\alias{get_url_from_opendataswiss}
\title{Get OWS URL from opendata.swiss}
\usage{
get_url_from_opendataswiss(dataset_id)
}
\arguments{
\item{dataset_id}{dataset_id from the catalog ( \code{\link[tgAPI]{getCatalog}})}
}
\value{
URL
}
\description{
Retrieves the OWS URL from opendata.swiss
}
