% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usage.R
\name{get_usage}
\alias{get_usage}
\title{Get Usage Data (only for internal usage with API Key)}
\usage{
get_usage(key, start, end)
}
\arguments{
\item{key}{API key from OpenDataSoft}

\item{start}{a date as character string in the format "YYYY-MM-DD" indicating the start of the date range in which the analytics data should be downloaded}

\item{end}{a date as character string in the format "YYYY-MM-DD" indicating the end of the date range in which the analytics data should be downloaded}
}
\value{
dataframe containing the usage analytics data.
}
\description{
Function to retrieve usage data from the backend. This function is for internal use only. An API key must be specified and passed to the function.
}
