% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wfs.R
\name{get_wfs}
\alias{get_wfs}
\title{Get WFSClient}
\usage{
get_wfs(dataset_id)
}
\arguments{
\item{dataset_id}{dataset_id from the catalog (see \code{\link[tgAPI]{getCatalog}})}
}
\value{
a ows4R WFS Environment ()
}
\description{
Get the WFSCLient for given dataset_id by using ows4R
}
\seealso{
\code{\link[ows4R]{WFSClient}}
}
